Expose param default values to hack (gated)
Summary:
Include param default values in the output if the include_assignments config is enabled.

In order to be able to "snip" the arbitrary initializer values without parsing, we need the "end location" of each parameter. To get that location, we add a "parameter_end" to the schema for ParameterDeclaration: it is only relevant for snipping initializer expressions.

Reviewed By: nt591

Differential Revision: D52660759

fbshipit-source-id: a221f5a278fb00fb4068974172ebdaef9d766705