Decl diffing: detect asyncification
Summary:
Decl diffing happens when computing fanout. It compute a semi-precise diff as a `ClassDiff.t`.

For method diffing, ClassDiff.t allows to say whether a method was added, removed or modified (+ some other stuff). We extend that to provide more info in case the method was modified:
```
type member_change =
  | Added
  | Removed
  | Changed_inheritance (* Modified in a way that affects inheritance *)
  | Modified of MemberModification.t
    (* Modified in a way that does not affect inheritance *)
  | Private_change_not_in_trait (* Added/removed a private member *)
```
with for now `MemberModification` being a simple variant covering just what we need to know for now:
```
module MemberModification = struct
  type t =
    | Asyncified
    | Deasyncified
    | Other
end
```

(Note that because of `Changed_inheritance` and `Private_change_not_in_trait`, we will fail to log asyncification which:
- are done on private methods
- are done on methods where we also change visibility or abstractness (causing `Changed_inheritance`)

It's fixable, but is it worth fixing?)

We don't use that new information for fanout computation, only for logging.

Reviewed By: enetsee

Differential Revision: D52656773

fbshipit-source-id: 16d0d19ae59e48d79788d3739e144cd3641c414b