Revert D52695743: Multisect successfully blamed "D52695743: [require class] update public documentation now that require class is no longer experimental" for otest failure
Summary:
This diff is reverting D52695743
D52695743: [require class] update public documentation now that require class is no longer experimental by francesco-zappa-nardelli has been identified to be causing the following test failure:

Tests affected:
- [hphp/hack/test/typecheck:extracted_from_manual_updated - main](https://www.internalfb.com/intern/test/844425023351611/)

Here's the Multisect link:
https://www.internalfb.com/multisect/3921963
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: dlreeves, francesco-zappa-nardelli

Differential Revision: D52704842

fbshipit-source-id: dd1f3c7cf46c5c93a92bed27c3f0cef9a04650c8