UserAttributeParams tests to show collections are dropped by the
Summary:
Title says it all. This is a limitation of the decl parser, but I'd
like tests that will fail/regress when we fix it. I like tests as
documentation, and this is at least meaningful to have lying around.

I intend to hopefully follow up with a fix eventually (on the order of this
quarter, not this week) unless someone else tackles it first.

Reviewed By: edwinsmith

Differential Revision: D52660931

fbshipit-source-id: 5a10797501b3a95b6aa4588af530ab61685c8cd5