Fix environment in XHP typechecking
Summary:
XHP literals are typechecked twice because we check it as a constructor call (for soundness) but then separately check against the attribute specification.

This causes refinements to be introduced and invalidated leading to a typehole, inaccurate types in the TAST, and incompleteness while typechecking.

This diff fixes that by restoring the local variable environment back to the state before the first typecheck when we are doing the second typecheck for attributes.

Reviewed By: andrewjkennedy

Differential Revision: D44241919

fbshipit-source-id: d7abda8636b631209c674b23b8378073a737b415