Add a way to set crate name without enforcing it matches namespace rust
Summary:
The `thrift_library` macro implements 2 ways to determine a crate name for the Rust crate:

```
# crate name is `record_if` based on the target name
thrift_library(
    name = "record-if",
    languages = ["rust"],
    thrift_srcs = {"record.thrift": []},
)

# crate name is `r_e_c_o_r_d` as explicitly set
thrift_library(
    name = "record-if",
    languages = ["rust"],
    rust_crate_name = "r_e_c_o_r_d",
    thrift_srcs = {"record.thrift": []},
)
```

As we are introducing support for `namespace rust` to set a crate name based on the contents of a Thrift source file, the Thrift code generator contains code to enforce that `namespace rust` determined from the source file isn't out of sync with the crate name determined from the Buck target.

```
// okay
namespace rust r_e_c_o_r_d

// not okay
namespace rust something.else
```

But in preparation for D44933146, there needs to be a way for this enforcement to occur only in the case of explicit `rust_crate_name`, not in the other case. In the other case, "record_if" will get used as a default name if there is no `namespace rust` in the source file, but if there is `namespace rust`, that will take effect and it doesn't matter whether it is equal to "record_if".

Reviewed By: diliop

Differential Revision: D52757885

fbshipit-source-id: 10c4e12c91ac81306e8dbabcd29a94503b9b53fb