Always resolve HH\classname to annot_type, tighten the types of class_to_classname
Summary:
Changes the behavior of the `RO::EvalClassPassesClassname` flag.

- **Previously:** when false, it would leave `HH\classname` as a normal, unresolved type alias to later be resolved to `string`, which would then optionally raise a notice about class-to-str conversion.

- **Now:** we unconditionally resolve it to `AnnotType::Classname`, and then defer the `RO::EvalClassPassesClassname` checks to when we're checking that annot type. Only the `ClassnameNoticesSampleRate` flag can affect the behavior of HH\classname, and to do a string conversion you must go through an explicit `string` type hint or string position (array key, etc).

This should slightly improve the flexibility of the class pointer migration by separating out the to-string sampling from the type hint behavior, and it lets me tighten the types on the explicit conversion builtin.

 ---

V1 summary

Talked with jano about his changes to native function enforcement, conclusion is just to tighten the signature but leave the impl to TypedValues for now because of the union on the input. Can move to String later if we start only accepting class inputs, since right now it's simpler to just return the input typedvalue

Reviewed By: jano

Differential Revision: D52707092

fbshipit-source-id: b218fccc6cd7fa0ae7fb5521d2edc3e947575259