Don't emit branches for module checks if cls/func is known
Summary:
Noticed this when refactoring {D52325661}, where I check `cls->hasConstVal()` for a different notice. If the SSATmp we give to the module check function has a const val, we don't need to go through the trouble of emitting these branches

https://www.internalfb.com/code/fbsource/[547b8fe9bb95a49a49f52262fab8073f5595ca79]/fbcode/hphp/runtime/vm/jit/irgen-call.cpp?lines=1355-1370

when we could just do the "known" checks directly.

Reviewed By: ricklavoie

Differential Revision: D52712383

fbshipit-source-id: dd57bbce6e93ee584e0727e8d0d5083dcae633e8