Always inline test method
Summary:
`StackTraceTest.AsyncStackTraceSimple` should behave identically despite what compiler flags (inline or noinline) were passed to build it.
We enforcing that via source code

`verifyAsyncStackTraces` is a test helper to assert the stack has specific frames stacked into it.
If we disable inlining, `verifyAsyncStackTraces` is stopped being inlined, so itself would show in the stack on top of `funcF()` frame. To fix it we force verifyAsyncStackTraces  inlining in code, and the stack goes back to how the test expects it (with nothing on top of funcF) and the tests pass again.

Since funcF (and others) itself has explicit inlining directives to have a predictable stack layout, it is right to force inline `verifyAsyncStackTraces`

Reviewed By: jbardini

Differential Revision: D52836847

fbshipit-source-id: c0b65d05dd46a4ec28e9e2b2e8478433ffdda3ab