Fix stepping-over lines with multiple calls
Summary: Step-over in the debugger is implemented as a step-in followed by a step-out. There can be multiple function calls in the same line the user is stepping over. In such a case, the step-out state will be `StepOutState::Out` after returning from the first call. It needs to be changed to `StepOutState::Stepping` when entering subsequent calls. Otherwise, the execution will pause inside the second function.

Reviewed By: paulbiss

Differential Revision: D52482394

fbshipit-source-id: 561a9ab7e4f2786a0058ad5cddeb8a4c944c9fde