Improve JIT for classname_to_class by using IR op for __DynamicallyReferenced notice
Summary:
Extracting this change from D52325661 and D52365879 because it's independently useful.
 ---
My optimization for classname_to_class was unnecessarily pessimistic; I was trying to pull the class name out of the clsTmp to make a string for RaiseNotice, which just gives too many opportunities for JIT to fail. Instead, add an op for this specific notice and pass the tmp directly, fetching the name at runtime.

Reviewed By: ricklavoie

Differential Revision: D52714374

fbshipit-source-id: b4c09d94f1c4c31be14f888bcf76d73e03142997