Remove extra-data from RaiseStrToClass IR op
Summary:
Following a tip from binliu19 on D52714374, I went back and revisited the discussion on D50202554; I think it was unnecessary to propagate the sample rate to this IR op; only `RaiseNotice` needs it since it raises heterogeneous notices.

Removes sample rate parameter and directly reads the runtime option in the helper.

Reviewed By: ricklavoie

Differential Revision: D52825943

fbshipit-source-id: 511f71de50f5793a7358d8f2c652a8850fd479c8