Cleanup usage of AnyArray parameter types in extensions
Summary:
AnyArray parameter type for builtins is currently unenforced, effectively
acting as mixed.

Let's convert all builtins that use it either to mixed to reflect the actual
situation, or in the case of ext_async_mysql to vec, since that API receives
only vecs.

The only one that was not updated is dummy_arraylike_builtin(), as it is used
only by HHVM unit tests. These unit tests will be updated when the enforcement
of parameter types for builtins is unified with non-builtins.

Differential Revision: D52710477

fbshipit-source-id: f833666b5f4fbcaa4bb55014aeff6d62c2ba9a7a