passing polymorphic allocator to AllocatingParserStrategy
Summary:
passing polymorphic_allocator to AllocatingParserStrategy will give us great flexibility in choosing how to manage memory.

As a complication, on mac, `memory_resource` header is placed under experimental namespace. For now, this feature is not supported on the mac, we can potentially add this support in the future when demand arises.

Reviewed By: robertroeser

Differential Revision: D52842006

fbshipit-source-id: 7492225b34b6fdf5dab54d72ecb8530782d2300f