Use logging comparators for RepoAutoloadMap
Summary:
We need type and function collision logging for RepoAutoloadMap, this diff adds it.

RepoAutoloadMap is backed by several Blob::HashMapIndex instances, and
HashMapIndex is parameterized by a CaseSensitive boolean flag. The actual
comparisons are defined in util/blob-writer.h and not exposed. In general, we don't
want anything in hphp/util to need to know or care about Hack language semantics.
This separation of concerns required some refactoring.

1. Change HashMapIndex to take a KeyCompare functor struct with equals() and hash()
methods. This was easy because the existing CaseSensitive flag was only used to
choose between two such structs, already.

2. Declare several KeyCompare structs for RepoFile: CaseSensitiveCompare, TypeNameCompare,
and FuncNameCompare, along with convenience wrappers for HashMapIndex<KC>. And
do the same thing for VirtualFileSystem, which needs a case sensitive PathCompare.

3. Factored tstrcmp_slice() and other logging helpers out of string-data-inl.h to isame-log.h,
so we can include them in repo-file.h without dragging in all of string-data.h.

doing it this way improves the codebase:
* only RepoFile and RepoAutoloadMap know about type and function names
* blob-writer.h is more flexible; more specialized functors can be introduced.

There's also an opportunity to change these functors to take slices instead of `const std::string&`
refs, which would avoid some string allocations, but this diff doesn't do it.

Reviewed By: ricklavoie

Differential Revision: D52834729

fbshipit-source-id: db20408dad323bf716926139820bcc087552b500