Clean up and document hackc-translator bits
Summary:
This came up and took me a minute to remember what exactly was going on here. Added some comments to make it more clear when this is revisited.

Also, it looks like we don't actually need the Dummy `_0` for `SSwitch`. I previously thought we needed it due to this error P1053800813, where by default, opcode inputs are labeled `_0`,` _1`, ...
For `SSwitch`, we rename these inputs at `targets` and `cases` in HackC so we pass in a dummy `_0` so the macro would succeed in passing in `body._0` to `handleSLA()`, even though it wouldn't actually read it.
When testing this, I realize since we don't read it, we don't evaluate it since the preprocessor just deletes the string `body._0` passed in.

```
#define handleSLA(ts, imm) do {                           \
  auto const targets = body.targets;                      \
  auto const cases = body.cases;                          \
  auto const imm1 = imm; \\          <--------------------------- error! P1053800813
  ts.fe->emitIVA(cases.len);                              \
  for (int i = 0; i < cases.len-1; i++) {                 \
    auto const caseName = toStaticString(cases.data[i]);  \
    auto const caseId = ts.mergeLitstr(caseName);         \
    ts.fe->emitInt32(caseId);                             \
    ts.addLabelJump(targets.data[i]);                     \
  }                                                       \
  ts.fe->emitInt32(-1);                                   \
  ts.addLabelJump(targets.data[cases.len-1]);             \
} while (0)
```

Reviewed By: mdko

Differential Revision: D52856465

fbshipit-source-id: a2ae3b89fc010effe7bb5f2de2c37bd580bce76a