Create hhvm_configtool
Summary:
Was messing around 1 hour later had this thing working so figured lets commit this

Next steps would be this can be an easy jump point to materialize the tier overrides, since the rules are straight forward with an import list of environment flags ie tiers, tags, etc

This output alone can help understand how changes to config.hdf works, since it de-nests the nesting into a flat KV pair

The unique_ptr stuff is messing with the tree being deconstructed while traversing from the root node, so added a special function to enumerate the nodes without losing the root children nodes.  If someone wants to fix that go for it, and eliminate the added workaround.

Reviewed By: edwinsmith

Differential Revision: D52863288

fbshipit-source-id: 8685d5aea9cae07ca67bd231da3f2070d9da7265