event logging for various interesting things in ext_facts
Summary:
I want to start collecting telemetry on all the weird corner cases and failure modes
that can happen with Facts. This is a start. Rare events bypass PerfSampleRate when
it's nonzero but still skip logging when it is zero.

Refactored more of the hhvm_sandbox_events logging code to a central location
runtime/base/sandbox-events.{cpp,h} since we need to call it from outside FactsLogger.

Reviewed By: nt591

Differential Revision: D52580966

fbshipit-source-id: 941187d990a0eb4071045d31378fe64f93e2fb8f