Fix forward for ManualTimekeeper deadlock
Summary:
This is a fix for D52635320 which broke a test (T175438010)

Delos' AsyncOperationSchedulerTest.operationTimeout saw a deadlock with the new code.  When ManualTimekeeper went to set the value on the future, it triggered a series of callbacks and a cancellation.  In essence, ManualTimekeeper grabbed a write lock in `advance()` and during callback processing called into the future interrupt handler, which deadlocked trying to acquire the lock again.

This change fixes the issue - rather than trying to dig through the schedule, the interrupt handler instead takes a reference directly to the promise to set the exception.

The main downside is that we no longer removed cancelled futures from the schedule on interrupt, but this is only used in tests so schedule bloat shouldn't be an issue.

Reviewed By: alpinejoe

Differential Revision: D52915051

fbshipit-source-id: f562c416b0be1c38f4805a69b8b8096e235508ec