Enable standard builtin param type enforcement in non-zend tests
Summary:
Set CheckBuiltinParamTypeHints to 2 in all tests except of zend, then update
tests as needed.

The main complication is that the special builtin parameter type enforcement
was throwing an exception, while the standard enforcement raises an error. This
required installing an error handler that rethrows them as exceptions (unless
there was only one error at the end of execution, in which case the expected
output was updated).

Tests where the enforcement behavior of two systems differs were converted to
soft enforcement (CheckBuiltinParamTypeHints=1) to demonstrate the behavioral
difference.

Other than that, the expected output was updated to the standard parameter type
enforcement error strings.

zend tests are kept at unenforced state for now, they will require additional
sweep (~500 of them failing).

Reviewed By: ricklavoie

Differential Revision: D52977813

fbshipit-source-id: b2e96ffc335662269e8188e4952bd5a63c3046d7