remove LSP implementation of class explorer features
Summary:
Remove the language server side of Hack Class Explorer because it is dead code.

Hack Class Explorer was removed from our vscode extension in D52046043 and the change was fully rolled out, so this diff is safe.

There are likely to be no other consumers, since we didn't use the standard LSP protocol for typeHierarchyProvider.

Reviewed By: patriciamckenzie

Differential Revision: D52944897

fbshipit-source-id: f0d75198b4b951203764d8dbfe6f4b76814a8a45