Remove extra-data from RaiseStrToClass IR op [take 2]
Summary:
original: D52825943; missed an occurrence of RaiseStrToClsNotice

Following a tip from binliu19 on D52714374, I went back and revisited the discussion on D50202554; I think it was unnecessary to propagate the sample rate to this IR op; only `RaiseNotice` needs it since it raises heterogeneous notices.

Removes sample rate parameter and directly reads the runtime option in the helper.

Reviewed By: binliu19, kavoor

Differential Revision: D52882317

fbshipit-source-id: c5413ecd9d0575bac9c3adb30aee7f4c5297e07e