tweak num reps in SpinLock.Correctness test
Summary:
Individual runs may take OTOO minutes and may hit timeouts in CI when loaded.

Since this is testing an exclusive mutex, the time taken scales with the number of threads and the lengths of the critical sections.

When choosing the number of reps per thread, take account of the number of threads and also take account of whether TSAN is enabled to try to hold the run time to OTOO seconds rather than OTOO minutes.

Reviewed By: Gownta

Differential Revision: D52885334

fbshipit-source-id: b2771cc121ed23ef56ca1df46935b07049123371