Do not reduce FCallClsMethodM to FCallClsMethodD when $1 is a string
Summary:
In HHBBC, `FCallClsMethodM` (e.g. `$x::foo()`) is reduced to `FCallClsMethodD` (e.g. `C::foo()`) when the LHS class is known and certain conditions pass e.g. modularity and a calculation about whether we care about a "dynamic call". The latter calculation is focused on method-level dynamism e.g. `__DynamicallyCallable`, so if `rfunc.mightCareAboutDynCalls()` returned false, the bytecode was still reduced, even if the LHS might be a string.

This change makes LHS string always be checked when the str-to-class notice flag is on.

Reviewed By: ricklavoie

Differential Revision: D52898731

fbshipit-source-id: a96b6b8eb18dde89d00d77beaab6e009301cdd77