InstrSeq::to_vec
Summary:
hackc needs to be able to transform function bodies with control-flow
based passes after the body is generated. To do so it needs to start with
a flattened Vec<Instruct> that is not arena-allocated.

Rename InstrSeq::compact() -> to_slice(), then add IsntrSeq::to_vec(),
sharing the underlying compaction code that squeezes out consequtive SrcLoc.

Reviewed By: kavoor

Differential Revision: D53001364

fbshipit-source-id: b60a24d09e55516fb6f3939b1ec29270dbc348e7