Revert D52863003: Multisect successfully blamed "D52863003: Register underlying struct in TypeRegistry" for otest failure
Summary:
This diff is reverting D52863003
D52863003: Register underlying struct in TypeRegistry by rmakheja has been identified to be causing the following test failure:

Tests affected:
- [thrift/lib/cpp2/op:std-serializer-test - StdSerializerTest.Struct](https://www.internalfb.com/intern/test/281475052388457/)

Here's the Multisect link:
https://www.internalfb.com/multisect/4085006
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Reviewed By: rmakheja

Differential Revision: D52983218

fbshipit-source-id: 06e13a86c5c7abfbc3c11c45953dd58587966443