kIsOptimize, kIsOptimizeSize
Summary:
On platforms (gcc, clang) which have the ability to report whether compiler optimizations are enabled, `kIsOptimize` is true when optimizations are enabled for speed or size and `kIsOptimizeSize` is true when optimizations are enabled for size.

`kIsOptimize` is intended for use particularly in stress tests, where we might detect whether optimizations are enabled to decide whether to run more reps, threads, etc.

`kIsOptimizeSize` is intended for use particularly when deciding whether to optimize widely-inlined or widely-instantiated code for speed or for size, in those cases where doing both at once is not possible.

Reviewed By: Gownta

Differential Revision: D52885541

fbshipit-source-id: c45b72be751dd6baaaff7fc7b4667f1c6e140508