Do not read `cntvct_el0` on mobiles
Summary:
Turns out it is not always legal to read `cntvct_el0` register on all
AARCH64, we got `EXC_BAD_INSTRUCTION` crashes on some older IPhones (Phone 5S,
iPhone 6+ and iPhone 6) on iOS 12 and 13.

There are couple of reports about similar problem in the Internet as well:
1. [tensorflow][1],
2. [easy_profiler][2].

To be on the safe side, I disabled read of `cntvct_el0` on all mobile devices
as we have limited testing abilities there. On mobile we will fallback to
`std::chrono::steady_clock`, which should be portable enough.

Reviewed By: ot

Differential Revision: D53041946

fbshipit-source-id: 6c97ff55a6d8e512e620e7fddca1774aa1c30d4a