add support for setting headers via context
Summary:
In this diff we add a way to write headers via context
This is for the use case of writing the kCat to the header in the follow up diff D53042701

We can AddHeaders to context with the new AddHeader method
This context can be passed to the generated thrift RPC ChannelClient methods
These generated methods calls SerialChannel.Call with context in the thrift library
SerialChannel.Call calls SerialChannel.sendMsg with context
SerialChannel.sendMsg adds the headers to the transport using setHeaders
When request.Write is called it calls writeVarHeader in header.go
writeVarHeader calls writeInfoHeaders
writeInfoHeaders writes both the persistent and normal headers to the transport

It has been tested before that the way persistentHeaders are written is a valid way to write headers that can be picked up by a C++ thrift client that is looking for the kCat header.
Since the headers are written in the same way as the persistentHeaders we are confident that this will work.

At the end of SerialChannel.sendMsg we call Flush
The normal headers are cleared in the Flush function so that they are only written for the single request and not on every request like the persistent headers.

Reviewed By: slasher-, echistyakov

Differential Revision: D53042527

fbshipit-source-id: a415f6377b48a6d7d44c9bdfd5c2e0d1a490897f