Support customizing signals
Summary:
This mimics the `SEGFAULT_SIGNALS` env var support from
the original `libsegfault.so`:
https://github.com/lattera/glibc/blob/master/debug/segfault.c#L148-L199.
Unlike glibc's `libsegfault.so`, which only watches `SIGSEGV` by default
we install handlers for all "fatal" signals (except `SIGTERM`).

The minor change this makes in behavior for `libFollySegFault.so` is
that we no longer install a handler for `SIGTERM`.

Reviewed By: luciang

Differential Revision: D53049261

fbshipit-source-id: dacc15efc19dd6a918cc5f1d8adf17609d1f3c74