Mark * nodes as is_wildcard
Summary:
We internally set the name for wildcard nodes to a count of *'s seen.  We should track if the node has a wildcard name so we can properly handle it when we render the name.

If we render the internal number it is incorrect since if we start to rely on the actual index we may run into name collisions and overwrite values in hard to debug ways.  Rendering it as a wildcard makes it clear we can never depend on the number index for configuration.

Reviewed By: edwinsmith

Differential Revision: D53114464

fbshipit-source-id: dbd2e67630b76c24c4f16569428223fc16ceea0c