Omit client and server modules if no services declared in Buck target
Summary:
In {D53166614} and {D53166618}, we need to make either of 2 choices:

1. **Every** `thrift_library` gets a clients and services crate, regardless of whether the target contains any service definitions

2. The clients and services crates only exist if the target defines nonzero services

As far as I know, it's impossible to do #2 based on whether the target *actually* contains services in its Thrift source code, only based on whether the `thrift_library` macro call lists any service names.

https://www.internalfb.com/code/fbsource/[0fa600724c20b87b38d536611b742c8287322f75]/fbcode/zippydb/if/TARGETS?lines=9-16%2C32-34%2C46

So far for Rust we have not paid attention to the listed service names in the Buck target, but the other languages look at it for a similar purpose as this.

On the basis that only 1252 out of 6042 Thrift targets in fbcode today contain services, I think limiting the clients and services crates is the right call to keep the Buck graph small. We can revisit down the line if there is any reason this turns out to be annoying.

This diff eliminates the `client` and `server` modules of Thrift library targets that do not contain any services according to Buck. Previously these used to be either empty modules, or actually nonempty modules if there were services defined in the Thrift source code but not declared in the Buck target.

Based on check_all.sh, there are zero occurrences of Rust code importing a service from a Thrift library for which the Buck target does not declare any services.

Reviewed By: zertosh, shayne-fletcher

Differential Revision: D53189031

fbshipit-source-id: e4fa7ffee4ead76802bebbc92ad514ff1d24d3b2