Merge ClsGetC and ResolveClass implementations
Summary:
The JIT implementation for ClassGetC does most of its work in a function `ldCls` that takes an SSATmp which is fetched from the top of the stack; I replaced `ResolveCls`'s implementation with this function.

This enables us to ensure that ResolveClass and ClassGetC have convergent behavior so that we can replace some occurrences of ClassGetC e.g. D52899768 for static props + that loading `C::class` as class pointers has the same behavior as ClassGetC on lazy-classes (excl. module checks).

Added new LdClsFallback case to cover the difference between ResolveClass and ClassGetC's fatal error strings.

> Note: the loss of the conditional module check in jit doesn't seem to affect any tests, and indeed the bytecode interpreter impl of ResolveClass does not do a module check at all: see Class::load vs. Class::resolve. I've an unconditional check on ResolveClass in child diff {D52899488}.

Reviewed By: ricklavoie

Differential Revision: D52921885

fbshipit-source-id: 8dbf51c8f8d28695379ef97a3926910c3c29635d