Remove special casing of coeffects for virtualized expression tree expressions
Summary:
The virtualized expression in expression trees can require coeffects (e.g., write_props and globals for client property assignments and calls to xhp class constructors). Expression trees can appear in pure functions, and since the virtualized expression is never run, its effects can be ignored. This was accomplished with a flag to disable coeffect checking inside of virtualized expressions.

This diff removes the flag, so that coeffect checking is uniform. The whole virtualized expression got wrapped in a 0-arg Efun, that was then immediately called. By omitting the call to the Efun, and giving it the `defaults` coeffect, we allow the virtualized expression to have coeffects inside the Efun which are not required on the enclosing context.

Reviewed By: andrewjkennedy

Differential Revision: D53047818

fbshipit-source-id: c1c24bd869256937c15c4b3bcacb67b45af9cf40