Add query() functions that take a callback
Summary:
When using the low-level api (i.e. `beginQuery()`) it was possible to pass in callbacks, but when using the higher level functions this was not supported.  There is no obvious reason why callbacks can't be supported in the higher level API.

Add support for query callbacks into `query()` and `querySemiFuture()` and there variations.

Reviewed By: fadimounir

Differential Revision: D52552982

fbshipit-source-id: 37ec7d91e4f59aa2fc2f4b2cdf23f514b8d5fba6