Extract F14 hashing utilities into detail namespace
Summary:
This diff moves many of the hashing utilities out of the `Policy` and `F14Table` classes so that they are more easily accessible; no longer always requiring a full `Policy` template parameter.

This change is a requirement for the future 'add wrapper for F14HashToken' diff.

It is worth noting that `splitHash` now uses the globally scoped `ShouldAssume32BitHash` and `IsAvalanchingHasher` templated value structs and not the corresponding `shouldAssume32BitHash` and `isAvalanchingHasher` methods on `BasePolicy`. There is the possibility for this being a breaking change if overrides of those methods are relied on. However, the tests indicate this is not a concern within folly and, as these are properties of the hasher rather than the policy, I think it unlikely it will be a problem elsewhere.

Reviewed By: Gownta

Differential Revision: D52959302

fbshipit-source-id: bfbf0a84bb818fcbb2d83215f5b2e5a79f2f1a0d