Modernize MySQL C++ access
Summary:
The admarket/customaudience/engagementaudiencebackfill code was using old MySQL access patterns.  The `beginQuery()` function - which returned a `QueryOperation` is lower level code and is only needed for some specialized handling.

Since this code was synchronously waiting for the results, switch it to use the `query()` function which does the same `run()` and `wait()` but handles it for the caller.

Note that some of this code was passing QueryOperations around.  Again this is low level stuff so I converted this to pass the Connection around instead.

Reviewed By: fadimounir

Differential Revision: D53061197

fbshipit-source-id: 1066e9248e49281e3112f9e7d279bd9c18c50737