Share the main poller thread across MuxIOThreadPoolExecutor instances
Summary:
If multiple `MuxIOThreadPoolExecutor`s exist in a process, currently each one has its own polling loop thread. This is inefficient because it misses out on the opportunity of batching `epoll` wake-ups across multiple executors.

This diff splits polling logic out of `MuxIOThreadPoolExecutor`, so it can be shared across multiple executors (and possibly other executor implementations as well). The new `EventBasePoller` is a singleton, so there is only one polling thread for the whole process. This should easily scale to a large number of `EventBase` hand-offs, and if eventually it doesn't, it can easily be sharded internally without changing the interface.

Reviewed By: dmm-fb

Differential Revision: D53163790

fbshipit-source-id: dd900927c4764405d3d191f13d3d1168b8778db8