Revert D53106330: Multisect successfully blamed "D53106330: Remove __EnableUnstableFeatures('modules') from tests" for otest failure
Summary:
This diff is reverting D53106330
D53106330: Remove __EnableUnstableFeatures('modules') from tests by milliechen has been identified to be causing the following test failure:

Tests affected:
- [hphp/hack/test/facebook/write_symbol_info:regression_tests - tests/cases/modules](https://www.internalfb.com/intern/test/281475098369256/)

Here's the Multisect link:
https://www.internalfb.com/multisect/4183045
Here are the tasks that are relevant to this breakage:

We're generating a revert to back out the changes in this diff, please note the backout may land if someone accepts it.

If you believe this diff has been generated in error you may Commandeer and Abandon it.

Differential Revision: D53254730

fbshipit-source-id: b0496a8ab3cf4c3e4163a7ba1831fba7f89614e8