Do not log handle exception in Thrift.
Summary:
Got paged into S392282 due to logspew of this line.

So when handler is throwing un-declared error, we took **unnecessary responsibility** and log it. Tw log are easily took over by this log: P1150453290.

For un-declared error, customer can see it on their client, for example: https://fburl.com/scuba/service_router/tgc8d9fy. So this log is not helpful.

Differential Revision: D53299043

fbshipit-source-id: 2c9ba3abda01eac798a012352c56e09dd3778956