Deserialize more structured representation for cratemap
Summary:
Preserve a separation between the crate name and the module name for multi-file Thrift libraries. This makes it possible for subcrate names to be constructed from it in {D53246158}.

Previously, for a crate-map like the following, which corresponds to a Buck target `:if` for the Thrift file `foo.thrift`:

```
foo if
```

we would deserialize it to the map `{"foo" => "::if_"}`.

This is problematic because while the top-level crate's name is `if_`, it becomes impossible to tell whether the `-types` subcrate is supposed to be `if_types` or `if__types`.

Reviewed By: zertosh

Differential Revision: D53297270

fbshipit-source-id: f9adb890a381faec9c1ff659f0809be6e47bfb67