Revert D53106253: Multisect successfully blamed "D53106253: Move Hack.Lang settings to configs.specification" for otest failure
Summary:
This diff reverts D53106253
D53106253: Move Hack.Lang settings to configs.specification by WizKid causes the following test failure:

Tests affected:
- [hphp/test:verify_slow_interp_jit_repo_retranslate-all_re - fbcode/hphp/test/slow/class_level_where_clauses/class_level_where_clauses_enabled.php](https://www.internalfb.com/intern/test/281475070220831/)

Here's the Multisect link:
https://www.internalfb.com/multisect/4198465
Here are the tasks that are relevant to this breakage:

The backout may land if someone accepts it.

If this diff has been generated in error, you can Commandeer and Abandon it.

Reviewed By: WizKid

Differential Revision: D53319353

fbshipit-source-id: 61ae0096495b6726347c5b9517bd2863250e85f5