Add option to supply debuginfo file to extract_from_core.py
Summary:
Adds option to supply hhvm.debuginfo file to `extract_from_core.py`.

I think the logic of determining which/where to get this filke should continue to live in `fetch_from_core.sh` -- I'll update that to properly handle `hhvm-tw` and other packages and pass in this flag when needed.

Reviewed By: vassilmladenov

Differential Revision: D53338313

fbshipit-source-id: 629f9690b93fefb529f6645b2320f25c65387f91