explicit include compiler warning
Summary:
This validator warns against "transitive" includes. These happen to work in Thrift C++, but they cause compilation issues in Thrift Rust and Thrift Python C-API. Even though transitive includes "work" in C++, there's still plenty of room for confusion or poorly defined behavior if there are two thrift files with the same module name!

If a struct or union has a field, then that field's type should refer to a named type either:
1. defined in the current program
2. defined in an explicitly `include`d thrift program.

Rather than fully resolving typedefs, we verify that each "hop" of the typedef is valid by visiting the immediate type with the same logic as struct fields.

This initial version does not cover `const` because it doesn't cause issues in as many contexts as transitively included field types.

Reviewed By: iahs

Differential Revision: D53295726

fbshipit-source-id: fa0afe208d79bb5e15342f6c8e0ef4fc3fc542c0