Support for recursive case types
Summary:
We want to add support for recursive case types (where something in the case type (like a shape) refers to the original alias:
```
case type C1 = int | shape("foo" => C1);
```
this is supported by adding a new TypeStructureKind indicating that a field is a recursive type (so it's not infinitely expanded).

The type structure for the above type would be:
```
dict [
  'kind' => 31, // TypeStructureKind::T_union
  'union_types' => vec [
    dict [
      'kind' => 1, // TypeStructureKind::T_int
    ],
    dict [
      'kind' => 14, // TypeStructureKind::T_shape
      'fields' => dict [
        'foo' => dict [
          'kind' => 32, // TypeStructureKind::T_recursiveUnion
          'case_type' => 'C1',
        ],
      ],
    ],
  ],
  'case_type' => 'C1',
]
```
Note that the typechecker does not yet support recursive case types so this is just in preparation of that.

Reviewed By: ricklavoie

Differential Revision: D50467691

fbshipit-source-id: df5bf6e3c8524f0e4cf9ba1e40bdea64b26633cb