Enforce non-safe method exclusion in suspended state
Summary:
While the loop is in suspended state, it is still "running", so other non-thread-safe methods should be excluded: both attempting to start a new loop or calling `checkIsInEventBaseThread()` should fatal.

We can rely on the existing assertions by considering a suspended loop as running in a fake thread; this can be represented by a second sentinel state to the loop thread id.

In order to do this we need to switch from using `std::thread::id` (which is an opaque type with a single sentinel value) to OS TIDs, which are non-negative numbers stored in a signed integer, so we can use any negative number for sentinels.

The cost of `getOSThreadID()` is comparable to `std::this_thread::get_id()` (they both use thread-local storage), and as an additional minor benefit debug assertions now print the TID instead of the pthread handle. TIDs are more useful in logs because they are also printed in log lines, so now it is possible to match them.

Reviewed By: Gownta

Differential Revision: D53333203

fbshipit-source-id: 9fbb43c44003ec522f71180dd989727612a39d1b