Add more tests for EventBase timeouts
Summary:
D52355843 fixed `EpollBackend` for the corner case where a handler is deleted by a timer that fires in the same loop, but a test should have caught the bug. Add such a test.

Also add a stress test for timers scheduler on evbs through executors, so we can check that integration works (especially with `MuxIOThreadPoolExecutor`).

Reviewed By: dmm-fb

Differential Revision: D53430071

fbshipit-source-id: 922f7c1d6ac9d49266811778a4efec417f35b82e