fix for cython 3.0.8 enum api.h bug
Summary:
This is my workaround for the reported bug https://github.com/cython/cython/issues/5973

Also Removed the use of "cdef public". cython_library is not setup to use public. As it is impossible to handling linking. If its needed for some feature we can use the static_cython_extension target which allows for public use.

Reviewed By: Filip-F

Differential Revision: D53442603

fbshipit-source-id: 1f68658174ae152ecd737a60cb623f4d4f06bc7b