use u32 for AdataId
Summary:
In HHAS, `AdataId` is an identifier that must always have the format `A_<n>`,
the the `A_` has no meaning, all we care about is `n`. Lets store `n` as a u32
and relegate the `A_` prefix cruft to just HHAS printing and parsing.

Since this just gets rid of one kind of string identifier we can land it now regardless
of what direction we take for interning other strings.

Differential Revision: D53156228

fbshipit-source-id: f3d87579793eed8ae7ad3cdf90b72cd5140fc8b8