handle filtering logic in facts
Summary: this is the part where we actually use the new filter structures for type flags.  i do all of the filtering in memory, rather than pushing the filtering down to the sqlite queries, which followed the pattern from kindfilter -- if we need to push the filtering down instead just let me know.

Reviewed By: edwinsmith

Differential Revision: D53311523

fbshipit-source-id: 7234720569d21fdab0124deb09baf08344b53daa