Use EventBase::loopWithSuspension() in MuxIOThreadPoolExecutor
Summary:
With `loopPoll()` we have no way to know if the evb could make progress without blocking, for example some event may have become ready while the loop callbacks were running. In that case we would send the evb back to the poller just to be immediately rescheduled.
Similarly, there is no way to know if there are pending keepalives, so we cannot correctly join the evb.
Also, some before-loop callbacks rely on being called just before blocking, so we need to replicate the same behavior as a normal blocking loop.

Switching to `loopWithSuspension()` all these both problems: now the loop returns only if it would otherwise block if `loop()` was called instead, so if it completes we know that there are no ready events and no pending keepalives. Resumption does not call the before-loop callbacks.

Finally, this makes time measurement more faithful to the behavior of `IOThreadPoolExecutor`: the returned idle time now accounts for the time the evb was not being driven, instead of being virtually 0 (because there is no blocking).

Reviewed By: dmm-fb

Differential Revision: D53329929

fbshipit-source-id: 571c522dd2182ab47369e4e92a826c1715db1e16