Sample filtering of recordings
Summary: Allow the decision to record a request to be sampled by both a rate and a filter. The filters take the form of `{Superglobal name}.{Dict key}={Expected value}`. For example, filtering on whether the request has POST data with key k and value v could be accomplished by adding the command line parameter `-vEval.RecordSampleRate=_REQUEST.k=v`. Multiple filters can be combined using `&`. If both a rate > 1 and a filter are set, then the rate coinflip is performed first, followed by the filtering. Thus, the rate serves as an upper bound on the amount of sampling performed.

Reviewed By: jtwarren

Differential Revision: D53066763

fbshipit-source-id: 3ecebf478b96305c26f935209e67807c4d596ecd