Fix hashing keys at initialization time
Summary:
Starting with hashbrown-0.13.2, the `hashbrown`'s `DefaultHashBuilder` initializes the keys the first time **a Hasher** (rather than the actual builder) is allocated, through a static variable.

This is dangerous, as indeed, this means the hashing key will be generated on first query/insert! If this happens in a forked-process, the keys will not be shared amongst processes!

This diff forces key initialization by using `ahash::RandomState` as the default hasher. This HashBuilder initializes its hashing key when calling `::new()` eagerly, rather than waiting for the first time an actual `Hasher` is built.

Reviewed By: dtolnay

Differential Revision: D53585605

fbshipit-source-id: 5d94b6dfade1ecda7c3efff826b482e21b8bacb3