Fix printing in cli-server xbox threads
Summary:
Currently when running with `Eval.UnixServerProxyXbox=1` stdout and stderr are available in the xbox thread but not written to by `echo` and other printers that use `ob_flush`. The issue is that we resolve whether or not the request is in CLI mode for the purpose of OB before we've initialized the new CLI server context.

Also fix `HH\execution_context()` to return `xbox` in xbox threads (mirroring the behavior of regular CLI usage of xbox) and stop swallowing fatals in xbox threads started from CLI server.

Reviewed By: edwinsmith

Differential Revision: D53594208

fbshipit-source-id: 6d75ec2e3ad4f17b3ce91d83a3efd05f104a12ed