Fix doc comment for cEscape
Summary:
I was using this to escape strings for some generated Python code, and
turns out, this function will escape `?` to `\?`!

I've never actually run into this escape sequence in practice, so it surprised
me. Seems like it used to be in the doc comment (D457649), but got lost when
that moved to `String.inl-h` (D428941), so I yoinked the comment from 2012.

Q: would it be worth adding a variant of this that does not escape question marks, since C++17 dropped support for trigraphs? That would be useful for my specific use case (codegen for Python/Rust/etc), but IDK about general interest.

Reviewed By: Orvid

Differential Revision: D53595397

fbshipit-source-id: 0ab3444bed07be9e90c592ebce4ef40ad3548f04