Make it easier to preserve block information when transforming blocks.
Summary:
Write writing new blocks they're often based on the existing block (adding an 'if' diamond to an existing block, for example). This is important for preserving the block exception information (`tcid`).

During translation the Infer backend wasn't preserving tcid when building constants (rewriting the entry block). When the entry block was the only reference to a catch block this was causing the catch block to be DCE'd.

Reviewed By: davidpichardie

Differential Revision: D53532623

fbshipit-source-id: cb7fdcf95439548e7f192996a95b1293e26155ac