dependency analysis for OCaml code built with Buck2
Summary:
This is the tool I've been using to find where to break dependencies in order to speed up our builds.
recently, I used this tool to make hh_client build twice as fast: D53042015
More context about the project: https://fb.workplace.com/groups/hacklang.eng/permalink/25704931645795441/

I'm committing the code to make it easier for teammates to experiment with building on it.

## Approach

There is a dependency graph of OCaml files. And there's a much-coarser dependency graph of buck targets.
Showing the two graphs together and restricting the view to only relevant edges can help us see where to cut dependencies.
That is what `--output dot` is for.

Also, sometimes removing unused dependencies helps. That's what `--output unused` is for. Note that `--output unused` is overly aggressive right now.

## Guide for reviewing

I recommend starting with README.md and the test plan.

## Potential follow-on work

If this tool continues to prove useful:
- potentially upstream something to autodeps tool to automatically forbid unused deps
- If this script proves useful enough and there is interest, we can see if there's interest in promoting it to fbcode//common/ocaml

Reviewed By: edwinsmith

Differential Revision: D52941074

fbshipit-source-id: 926957ead29cf123827674cb6037930036bc6eb5