Produce clients based on service names from Thrift, not from Buck target
Summary:
I need this in order to implement support for `rust.Name` on services.

In `thrift_library` Buck targets, we declare a list of the services in each Thrift source file, because some languages generate a different library per service (unlike Rust). However, the service names written there in the Buck target are not necessarily the Rust name for the service; it is Thrift's name for the service. Today those are always the same, but with the introduction of `rust.Name` annotation support on services in the next diff in this stack, they are not. The code generated by `clients_builder.py` needs to refer to services by their Rust name, not Thrift name.

Reviewed By: shayne-fletcher

Differential Revision: D53671407

fbshipit-source-id: 89069010ca74da7433054289f4287c4d79da2d45