Remove zend tests that just test for builtin parameter types, [1/?]
Summary:
There are plenty of zend tests that just try to pass all various kind of
garbage to various builtins.

These tests change their output (fataling at the first invalid param type) when
Eval.CheckBuiltinParamTypeHints is set to the desired value of 2. In this
configuration, there is nothing special about builtins and we already have a
good coverage for general parameter type enforcement.

There is not much value provided by these tests, so remove 50 of them. Either
completely, or the parts that test for incorrect param types.

Differential Revision: D53684552

fbshipit-source-id: a0815256b9f4045696b54cb651b37a66a78c2c48