Remove legacy check for unenforced hhi
Summary: Generally we treat signatures in `hhi` files the same way as ordinary Hack code, and assume the same "enforcement" of types in order to determine our Sound Dynamic checking and implicit pessimisation. But there was an issue with the legacy behaviour of *not* trusting hhi signatures that has leaked into property assignment behaviour. Let's kill the legacy behaviour.

Reviewed By: madgen

Differential Revision: D53665914

fbshipit-source-id: d9f659f61aae01a298d74a3fe7d348aa3384da2d