thrift: to stop relying on operator->
Summary:
C++20 supports iterators with no operator->
I would like to remove the `operator->` requirement from the diff.

## Reasoning.

We have a few compact maps we use for deserialization.

The `reference` type on a map like this is akin to `std::pair<int, std::string_view>`.

So the only way I know how to support the `operator->` on the iterator is with `arrow_proxy` trick.

Unfortunately, this tends to allow code that is fine for a normal map but for this one leads to crashes.
Example: https://godbolt.org/z/hnWfqxMj4

Reviewed By: vitaut

Differential Revision: D53665618

fbshipit-source-id: 6e74dcf241e2477350cf7edbf8352ec0d34e82a6