Use the correct diagnostics API in parse_and_mutate_program
Summary: It doesn't make sense to require users pass `diagnostic_context` to `parse_and_mutate_program` because the context is only relevant for visitation which is an implementation detail here and we shouldn't expose cache to callers. Replace it with a more appropriate diagnostics API, `diagnostics_engine`, and only introduce the cache when running mutators. Apart from being cleaner, this also prevents accidentally corrupting the cache when the context is reused.

Reviewed By: iahs

Differential Revision: D53623058

fbshipit-source-id: 3a80d4893eeb7ab18bdd2b00c3f48c8dbb4b2f52