sbe message wrapper
Summary:
Adds a wrapper class for SBE messages to make them eaiser to work with. SBE doesn't have the concept of embedded messages or sub structs like Thrift, but this adds convince methods that makes it easier to work with sub messages.

Additionally, SBE has a 'set' concept but that only applies only applies to fixed length types. The message wrapper has some methods to help work with a 'union' of messages - the getSubMessageTemplateIdAndBuffer method. You can use this method to get the template id of the message from a field to determine what fly to apply.

Has helper encoding and decoding methods that can wrap an IOBuf and std::string_view.

Reviewed By: iahs

Differential Revision: D53367478

fbshipit-source-id: 0f7aebd4b3458826dd4779b2c1f7538e15edfdd5