Add logging for hh_fanout code
Summary:
I think most of the hh_fanout code could be dead.

If so, I can delete so much of this (probably everything except the calculate method)!!!

This will also mean that I can delete some of the parts of naming_sqlite.ml and naming_table.ml that I'd otherwise need to update for my sort_text feature.

In this diff, I basically log when these methods are called.

If i see no logs for these methods, I can delete them safely. Ty for the suggestion, vassilmladenov and ljw1004.

Differential Revision: D53482558

fbshipit-source-id: 3929145ec050444607981aa4a518bb36c15433c9