Improve management of EventBase keepalives
Summary:
The `EventBase` keepalive count is split across two counters, one non-atomic for in-loop operations, and one atomic for external operations (which mostly come from futures).
This requires `keepAliveRelease()` to always wake up the loop, because that's the only place where it can be determined whether the counts add up to 0. The cost of doing this is very large, and more than offsets the atomic operation (which is expected to be uncontended if the majority of operations are in-loop). Furthermore, `inRunningEventBaseThread()` is not cheap either.

With this diff we just use a single atomic count, so that the loop can be awoken only when the last keepalive is released, which in the common case never happens as the evbs are run with `loopForever()` in IO executors.

Reviewed By: andriigrynenko

Differential Revision: D53385046

fbshipit-source-id: 4520cb8a5c7a1edd3f978650ccf0aebba754c712