Remove zend tests that just test for builtin parameter types, [2/?]
Summary:
There are plenty of zend tests that just try to pass all various kind of
garbage to various builtins.

These tests change their output (fataling at the first invalid param type) when
Eval.CheckBuiltinParamTypeHints is set to the desired value of 2. In this
configuration, there is nothing special about builtins and we already have a
good coverage for general parameter type enforcement.

There is not much value provided by these tests, so remove 50 of them. Either
completely, or the parts that test for incorrect param types.

Reviewed By: mdko

Differential Revision: D53822127

fbshipit-source-id: 307199cd53a404966c63c00cfe161501be0d712e