Shift modularity checks onto ResolveClass for object construction
Summary:
`ResolveClass` was originally made as a bytecode target for `C::class` {D19561707}; when a flag was set it would create a class pointer. Currently, it is only used for a (quite complex) reified generics stanza on instantiation
```
# new D()
  .try {
    .srcloc 4:8,4:14;
    ResolveClass "D"
    PopL _1
    NewObjD "D"
    PopL _3
    CGetL _1
    ClassHasReifiedGenerics
    JmpZ L0
    CGetL _1
    CheckClsRGSoft
    Jmp L1
  L0:
    CGetL _1
    HasReifiedParent
    JmpZ L1
    CGetL _3
    NullUninit
    NewCol Vector
    CastVec
    FCallObjMethodD <> 1 1 "" "" - "" "" NullThrows "86reifiedinit"
    PopC
    Jmp L1
  L1:
    CGetL _3
    Jmp L2
  } .catch {
    UnsetL _1
    UnsetL _2
    UnsetL _3
    Throw
  }
L2:
  UnsetL _1
  UnsetL _2
  UnsetL _3
  Dup
  NullUninit
  .srcloc 4:8,4:14;
  FCallCtor <LockWhileUnwinding> 0 1 "" "" - "" ""
  PopC
  LockObj
```
and we end up calling Class::load twice.

This changes the modularity calculation (Class::resolve) to happen during `ResolveClass` instead of `NewObjD` and unconditionally restores it for the JIT (see parent diff)

Differential Revision: D52899488

fbshipit-source-id: 316ce137642018a0fca5f98ede6c964b994ce9d7