Use ResolveClass instead of ClassGetC to compile known class refs, fixes incorrect str-to-class notices on C::$foo
Summary:
The code to compile known class refs like the `C` in `C::$foo` is suboptimal. It emits
```
String "C"
ClassGetC
```
which means the runtime emit a str-to-class conversion notice for `C`; in fact it cannot tell the case apart from `$c::$foo`.

```
# $c::$foo
  String "foo"
  CGetL $c
  ClassGetC
  CGetS Mutable
# C::$foo
  String "foo"
  String "C"
  ClassGetC
  CGetS Mutable
```

**Unfortunately modules relies on this behavior** as modularity checks are emitted for `ClassGetC` only when $1 is a string

> Aside: this means modularity checks are elided when $1 is a lazy class.

In {D52899488}, I made `ResolveClass` responsible for checking modules when the class pointer is loaded, and here I've made us directly resolve the class pointer when the user references a class:

```
# C::$foo
  String "foo"
  ResolveClass "C"
  CGetS Mutable
```

So regular property static access will no longer erroneously raise a notice about str-to-class conversion.

Reviewed By: viratyosin

Differential Revision: D52899768

fbshipit-source-id: c5941b7e3b56d76995aa27ea23d2dadc137173eb